// src/types/domain.ts
export enum Status {
  PENDING = 'Pendiente',
  READY = 'Listo',
  USING = 'En uso',
}
// ...
export type MachineType = 'Cabezal' | 'Casetera' | 'Freno';
// ...
export interface Movimiento {
  id: string;
  machine_id: string;
  machine_type: MachineType;
  fecha: string;
  tecnico_nombre: string;
  motivo: string;
  origen: string;
  destino: string;
}

export interface Cabezal {
  id: string;
  tipo: 'Cabezal';
  estado: Status;
  ubicacion: string;
  freno_actual_id?: string;
  historial_movimientos: Movimiento[];
}

export interface Casetera {
  id: number;
  tipo: 'Casetera';
  estado: Status;
  ubicacion: string;
  historial_movimientos: Movimiento[];
}

export interface Freno {
  id: string;
  tipo: 'Freno';
  fecha_inicio: string;
  estado: Status;
  cabezal_id?: string;
  ubicacion: string;
  historial_movimientos: Movimiento[];
}